notes: implement a method to merge diverging notes
To handle the situation when concurrent GitGitGadget operations want to
modify the state that is persisted in Git notes (such as
`refs/notes/gitgitgadget`), we need a way to reconcile diverging Git
notes.

Happily, the way GitGitGadget works makes this relatively
straight-forward, as concurrent operations usually do not conflict and
merely need to add/modify/remove independent attributes and/or values.
This allows us to treat these Git notes as kind of a Conflict-free
Replicated Data Type (https://en.wikipedia.org/wiki/CRDT).

The strategy implemented in this patch is relatively simple, and follows
the "local-first" idea (where state is applied locally first and only
then persisted remotely): The newly added method determines what changes
were made locally, and then reapplies them on top of the upstream note
commit. The local operations need to fall into one of the following
categories:

- a note has been appended to

- a single line of text (which may, or may not, be a stringified JSON
  object) is added as a note on an object that had not had a note yet

- a note containing a single line of text has been replaced with a new,
  single line of text

- a note that contained a JSON object (stringified as a single line of
  text) has been modified. Valid modifications include:

  - an attribute that was added

  - an attribute that has been removed

  - an attribute containing a primitive type (string, number or boolean)
    has been modified (if local and upstream diverged, local wins)

  - items from an attribute containing an array have been removed and/or
    added

  - an attribute containing a Plain Old Javascript Object has seen any
    of these modifications

Other modifications are as yet unhandled.

For historical reasons (*cough* Git commands implemented in Unix shell
scripts *cough*) it is impossible to add a Git note that does not end in
a newline character. Let's ensure that we don't do that, either.

To avoid tampering with a Git index, let's use a temporary index. This
requires the environment variable `GIT_INDEX_FILE` to be passed to the
Git processes, via the `env` attribute of the `IGitOptions` (which only
need to be declared and no other changes are needed because the
`options` are then used as dugite's `IGitExecutionOptions`, which
already has the `env` attribute).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>