git: allow operating on bare repositories
We actually do want to work exclusively on bare repositories. A recent
Git version requires either the `safe.bareRepository` config setting to
indicate that it is safe, or the git dir to be specified explicitly via
`--git-dir`. Let's do the latter.

This will become very important when we switch to using GitHub Actions
with ephemeral (and partial) clones of `gitgitgadget/git`.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>